Изменение : MpaDecFilter - небольшие корректировки битстрима TrueHD.
git-svn-id: svn://svn.code.sf.net/p/mpcbe/code/trunk@3909 7426f7ed-705a-47fc-8cda-d59144bb0dce

+++ src/filters/transform/MpaDecFilter/MpaDecFilter.h
@@ -102,11 +102,11 @@ protected:
 
 			int ratebits;
 
-			int prev_frametime;
+			uint16_t prev_frametime;
 			bool prev_frametime_valid;
 
-			int mat_framesize;
-			int prev_mat_framesize;
+			uint32_t mat_framesize;
+			uint32_t prev_mat_framesize;
 
 			DWORD padding;
 		} TrueHDMATState;
